+++ CHANGELOG.md
@@ -34,6 +34,8 @@
   descendants to another location.
 * `Route().move_to(new_url, *, move_children)` can be used to move a `Route`
   and (optionally) it's descendants to another location.
+* `Route().swap_with(new_url, *, move_children)` can be used to swap a `Route`
+  and (optionally) it's descendants with another `Route`.
 
 ### Changed